GH-1: Moved javascript source to .
