+++ demos/multi-passthrough.html
@@ -11,16 +11,16 @@
 
     <script src="../node_modules/infusion/dist/infusion-all.js"></script>
 
-    <script src="../src/core.js"></script>
-    <script src="../src/system.js"></script>
-    <script src="../src/messageEventHolders.js"></script>
-    <script src="../src/connection.js"></script>
-    <script src="../src/multiPortConnector.js"></script>
+    <script src="../src/js/core.js"></script>
+    <script src="../src/js/system.js"></script>
+    <script src="../src/js/messageEventHolders.js"></script>
+    <script src="../src/js/connection.js"></script>
+    <script src="../src/js/multiPortConnector.js"></script>
 
-    <script src="../src/ui/templateRenderer.js"></script>
-    <script src="../src/ui/multiSelectBox.js"></script>
-    <script src="../src/ui/multiPortSelectorView.js"></script>
-    <script src="../src/ui/svgIcon.js"></script>
+    <script src="../src/js/ui/templateRenderer.js"></script>
+    <script src="../src/js/ui/multiSelectBox.js"></script>
+    <script src="../src/js/ui/multiPortSelectorView.js"></script>
+    <script src="../src/js/ui/svgIcon.js"></script>
 
     <script src="./js/multiPassthrough.js"></script>
