+++ index.js
@@ -10,11 +10,11 @@
 "use strict";
 var fluid = require("infusion");
 
-require("./src/core");
-require("./src/messageEventHolders");
-require("./src/connection");
-require("./src/portConnector");
-require("./src/system");
+require("./src/js/core");
+require("./src/js/messageEventHolders");
+require("./src/js/connection");
+require("./src/js/portConnector");
+require("./src/js/system");
 
 // Register our content so it can be used with calls like fluid.module.resolvePath("%youme/path/to/content.js");
 fluid.module.register("youme", __dirname, require);
