+++ src/js/ui/svgIcon.js
