+++ src/js/ui/templateRenderer.js
