+++ tests/html/connection-tests.html
@@ -18,11 +18,11 @@
         <script src="/testem.js"></script>
         <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
 
-        <script src="../../src/core.js"></script>
-        <script src="../../src/messageEventHolders.js"></script>
-        <script src="../../src/connection.js"></script>
+        <script src="../../src/js/core.js"></script>
+        <script src="../../src/js/messageEventHolders.js"></script>
+        <script src="../../src/js/connection.js"></script>
 
-        <script src="../../src/test/webMidiMock.js"></script>
+        <script src="../../src/js/test/webMidiMock.js"></script>
         <script src="../js/connection-tests.js"></script>
     </head>
