+++ tests/html/findPorts-tests.html
@@ -18,7 +18,7 @@
         <script src="/testem.js"></script>
         <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
 
-        <script src="../../src/core.js"></script>
+        <script src="../../src/js/core.js"></script>
 
         <script src="../js/findPorts-tests.js"></script>
     </head>
