+++ tests/html/multiPortConnector-tests.html
@@ -18,13 +18,13 @@
         <script src="/testem.js"></script>
         <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
 
-        <script src="../../src/core.js"></script>
-        <script src="../../src/system.js"></script>
-        <script src="../../src/messageEventHolders.js"></script>
-        <script src="../../src/connection.js"></script>
-        <script src="../../src/multiPortConnector.js"></script>
+        <script src="../../src/js/core.js"></script>
+        <script src="../../src/js/system.js"></script>
+        <script src="../../src/js/messageEventHolders.js"></script>
+        <script src="../../src/js/connection.js"></script>
+        <script src="../../src/js/multiPortConnector.js"></script>
 
-        <script src="../../src/test/webMidiMock.js"></script>
+        <script src="../../src/js/test/webMidiMock.js"></script>
         <script src="../js/lib/fixtures.js"></script>
         <script src="../js/multiPortConnector-tests.js"></script>
     </head>
