+++ tests/html/system-tests.html
@@ -18,10 +18,10 @@
         <script src="/testem.js"></script>
         <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
 
-        <script src="../../src/core.js"></script>
-        <script src="../../src/system.js"></script>
+        <script src="../../src/js/core.js"></script>
+        <script src="../../src/js/system.js"></script>
 
-        <script src="../../src/test/webMidiMock.js"></script>
+        <script src="../../src/js/test/webMidiMock.js"></script>
         <script src="../js/lib/fixtures.js"></script>
 
         <script src="../js/system-tests.js"></script>
