+++ tests/html/webMidiMock-tests.html
@@ -18,7 +18,7 @@
         <script src="/testem.js"></script>
         <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
 
-        <script src="../../src/test/webMidiMock.js"></script>
+        <script src="../../src/js/test/webMidiMock.js"></script>
 
         <script src="../js/webMidiMock-tests.js"></script>
     </head>
